fix: pre-wrap small uints to avoid overflows

+++ contracts/libraries/QuotasLogic.sol
@@ -75,7 +75,7 @@ library QuotasLogic {
         pure
         returns (uint256)
     {
-        return quoted * (cumulativeIndexNow - cumulativeIndexLU) / RAY; // U: [QL-2]
+        return uint256(quoted) * (cumulativeIndexNow - cumulativeIndexLU) / RAY; // U: [QL-2]
     }
 
     /// @dev Calculates interest accrued on quota since last update
@@ -169,7 +169,7 @@ library QuotasLogic {
             // For some tokens, a one-time quota increase fee may be charged. This is a proxy for
             // trading fees for tokens with high volume but short position duration, in which
             // case trading fees are a more effective pricing policy than charging interest over time
-            caQuotaInterestChange += change * tokenQuotaParams.quotaIncreaseFee / PERCENTAGE_FACTOR; // U: [QL-3]
+            caQuotaInterestChange += uint256(change) * tokenQuotaParams.quotaIncreaseFee / PERCENTAGE_FACTOR; // U: [QL-3]
 
             // Quota revenue is a global sum of all quota interest received from all tokens and accounts
             // per year. It is used by the pool to effectively compute expected quota revenue with just one value
